Fix implicit numeric widening warnings in rspace-bench

+++ rspace-bench/src/test/scala/coop/rchain/rspace/bench/BasicBench.scala
@@ -161,7 +161,7 @@ object BasicBench {
       Arbitrary(
         for {
           v <- Gen.posNum[Int]
-        } yield GInt(v)
+        } yield GInt(v.toLong)
       )
 
     def onePar(i: GInt) = List(
